# Optimize the lookups of fields in the 'Header' and reduce the cost of allocations