# Description of Changes

Eliminate the linear scans in lookups for `FieldName`:

```bash
# Before
find_by_name            time:   [13.951 µs 14.554 µs 15.303 µs]
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high severe  

#After
find_by_name            time:   [269.77 ns 270.10 ns 270.51 ns]
                        change: [-98.148% -98.082% -98.027%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
```

Using a map of `FieldsName -> Postions`. 

NOTE: I don't use a regular `HashMap` or `Map<FieldName, Column>` because the order of the fields should be preserved as intended by the user (ie: `select c, a, b, i_want_this_order`).

Also, considering that the header is rarely updated, the `FieldNames` are now under `Arc`.

Closes #932.

# Expected complexity level and risk

1