@kazimuth 2024-03-07
Yeah, I was also wondering if it would be possible to switch entirely to integer identifiers for fields. (Also tables, but that would be later...)