@joshua-spacetime 2024-03-21
@mamcx since this is no longer critical for achieving our performance target, I suggest we close this and come back to it as part of refactoring the query engine.